Added utils function:(pure and bubbleSort)
