--- src/utils/reset.js
@@ -1,10 +1,0 @@
-export const resetRootFontSize = () => {
-  let deviceWidth = window.innerWidth
-  let devicePixelRatio = window.devicePixelRatio || window.webkitDevicePixelRatio || window.mozDevicePixelRatio
-  let calc = 7.5
-  if (deviceWidth > 414) deviceWidth = 414
-  if (deviceWidth < 320) deviceWidth = 320
-  if (deviceWidth < 320 && devicePixelRatio >= 2) calc = calc - (devicePixelRatio - 1)
-  // 计算设计稿和实际像素的缩放比。向上取整1px = 0.01rem
-  document.documentElement.style.fontSize = Math.ceil(deviceWidth / calc) + 'px'
-}
